v0.9.4 - (1.0b)
* final 1.0b release build

- started basic live streaming integration
- fixed error after some refactoring
- renamed exported interfaces to better understanding names

git-svn-id: file:///C/Users/thorsten/Dropbox/SVN/VideoDashboard/trunk@5 85b27f27-d72a-e640-9e06-118b1291f79f

+++ libDataLogging/ibtDataLogger.cpp
@@ -180,7 +180,7 @@ void CibtDataLogger::ParseSessionInfo(const char* sessionInfoStr)
 
 			//float t = stof(s.str().c_str());
 			//t = roundf((t * 1000) / 1000);
-			float t = roundf((atof(val) * 1000.0f));
+			float t = roundf(((float)atof(val) * 1000.0f));
 			t /= 1000.0f;
 
 			//t = (t * 1000) / 1000;
